Add Render and Transform Component + GameObject entity
Update ECS to last version
Start of View

+++ ECS/src/System/ISystem.cpp
@@ -7,7 +7,7 @@
 namespace ecs
 {
 
-  ISystem::ISystem(SYSTEM_PRIORITY priority, float updateInterval)
+  ISystem::ISystem(SYSTEM_PRIORITY priority, long updateInterval)
           : _updateInterval(updateInterval), _priority(priority)
   {
   }
